requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==4.4.2
emoji==0.5.4
Flask==1.1.1
future==0.18.2
huey==2.2.0
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mongoengine==0.19.1
numpy==1.18.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.5.2
pymongo==3.10.1
python-dateutil==2.8.1
python-gitlab==2.1.2
python-telegram-bot==12.5.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
